Write a C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
 /**
- * main - a program to print the size of functions with sizeof function
- * Return: 0 (success)
+ * main - a program to print the size of various computer types
+ * Return:0 (success)
 */
 
 int main(void)
 {
-	printf("size of a char: %u byte(s)\n", sizeof(char));
-	printf("size of a int: %u byte(s)\n", sizeof(int));
-	printf("size of a long int: %u byte(s)\n", sizeof(long int));
-	printf("size of a long long int: %u byte(s)\n", sizeof(long long int));
-	printf("size of a float: %u byte(s)\n", sizeof(float));
+	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("size of a int: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
